dev-postgresql: subtract the tool itself connection to connection query number
As the connection query will also count the connection made by this tool to make that query, and that this connection can be considered as "residual", it is subtracted from the connection count. For example, if the tool is configured for a threshold of 1 connection (which would be weird), the tool would always return 1 if its connection is not subtracted (which would be even more weird, since the database is not having a real usage).